Backend/requirements.txt
slack_sdk==3.41.0
slack_bolt==1.28.0
python-dotenv==1.2.2
qdrant-client==1.12.0
psycopg2-binary
asyncpg
playwright
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=12.2.0
groq
pymupdf
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
celery==5.4.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
sentence-transformers==2.7.0
torch>=2.0.0
httpx>=0.27.0
python-multipart==0.0.17
aiofiles==24.1.0
PyJWT>=2.8.0
email-validator>=2.0.0
tiktoken>=0.7.0
nltk>=3.8.0
numpy
supabase